+++ src/Vitorm/Sql/Transaction/ETransactionState.cs
@@ -1,0 +1,8 @@
+﻿namespace Vitorm.Sql.Transaction
+{
+    public enum ETransactionState
+    {
+        Active, Committed, RolledBack, Disposed
+    }
+
+}