+++ src/Vitorm/Transaction/ITransactionManager.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Vitorm.Transaction
+{
+    public interface ITransactionManager : IDisposable
+    {
+        ITransaction BeginTransaction();
+    }
+}